MVP-3 1. Правки ошибок скрипта от компилятора 2. Поправил package у темплейтов kt файлов 3. Почистил код
